wc - Moved AnalysisRegistry components to deprecated
