+++ src/webcomponents/variant/analysis/deprecated/test/knockout.20201029141213.SChLEA.js
@@ -375,4 +375,4 @@ export default [
             }
         ]
     }
-]
+];
