+++ src/webcomponents/variant/analysis/deprecated/test/knockout.20201103172343.kFIvpr.individuals.js
@@ -240,4 +240,4 @@ export default [
             }
         ]
     }
-]
+];
